Skeleton Outline for Code
Created a skeleton file which maps out some of the high level
functions that are going to be written. It is expected that when
writing for one of these functions that you divide necessary
tasks into smaller subfunctions if it gets too big. That
function should be defined right below the higher level
function. Currently each function takes no inputs and gives no
outputs, modify these as needed. Update comments so that it
accurately documents what each function does and how. DO NOT USE
TABS, DO NOT LEAVE TRAILING WHITESPACES.